Added persistent login state using Shared Preferences
