+++ lib/constants.dart
@@ -6,3 +6,7 @@ Color primaryGray = Colors.grey[100]!;
 const secondaryGray = Color(0xFF6E7191);
 const Color primaryRed = Color(0xFFE85D75);
 const Color bgColor = Color(0xFFF7F7F8);
+
+class Constants {
+  static String keyLoggedIn = "loggedIn";
+}
