+++ pubspec.yaml
@@ -44,6 +44,7 @@ dependencies:
   iconsax: ^0.0.8
   dart_mappable: ^4.2.2
   flutter_riverpod: ^2.5.1
+  shared_preferences: ^2.2.3
 
 dev_dependencies:
   flutter_test:
